automation/requirements.txt
# Multi-Agent System Dependencies
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MongoDB Atlas Dependencies
pymongo>=4.6.0
motor>=3.3.0

# AI/ML Dependencies
openai>=1.0.0
numpy>=1.24.0

# Git Operations
gitpython>=3.1.0

# Environment Management
python-dotenv>=1.0.0

# System Monitoring Dependencies
psutil>=5.9.0

# Kafka Dependencies
kafka-python>=2.0.2
aiokafka>=0.8.0

# Vector Database Dependencies (MongoDB Atlas via pymongo)
sentence-transformers>=2.2.0

# Web Framework (for health checks and monitoring)
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0